+++ src/Notify/assets/src/controller.ts
@@ -52,13 +52,7 @@ export default class extends Controller {
 
         this.eventSources.forEach((eventSource) => {
             const listener = (event: MessageEvent) =>
-                this._notify(
-                    JSON.parse(event.data).summary,
-                    JSON.parse(event.data).tag,
-                    JSON.parse(event.data).body,
-                    JSON.parse(event.data).icon,
-                    JSON.parse(event.data).renotify
-                );
+                this._notify(JSON.parse(event.data).summary, JSON.parse(event.data).content);
 
             eventSource.addEventListener('message', listener);
             this.listeners.set(eventSource, listener);
@@ -80,20 +74,11 @@ export default class extends Controller {
         this.eventSources = [];
     }
 
-    _notify(
-        content: string | undefined,
-        tag: string | undefined,
-        body: string | undefined,
-        icon: string | undefined,
-        renotify: boolean | undefined
-    ) {
-        if (!content) return;
-        if (!tag) tag = '';
-        if (!icon) icon = '';
-        if (!body) body = '';
+    _notify(title: string | undefined, options: NotificationOptions | undefined) {
+        if (!title) return;
 
         if ('granted' === Notification.permission) {
-            new Notification(content, { tag: tag, body: body, icon: icon, renotify: renotify });
+            new Notification(title, options);
 
             return;
         }
@@ -101,7 +86,7 @@ export default class extends Controller {
         if ('denied' !== Notification.permission) {
             Notification.requestPermission().then((permission) => {
                 if ('granted' === permission) {
-                    new Notification(content, { tag: tag, body: body, icon: icon, renotify: renotify });
+                    new Notification(title, options);
                 }
             });
         }